requirements.txt
django>=5.0.1,<6.0.0 # 5.0.1 이상 6.0.0 미만
djangorestframework>=3.14.0,<4.0.0 # DRF
psycopg2-binary>=2.9.9,<3.0.0 # psycopg2 : PostgreSQL connector
drf-spectacular>=0.27.2,<0.28.0 # DRF-Spectacular
uwsgi>=2.0.24,<2.1.0 #uWSGI

# docker-compose build